refactor: 사용자 설정 변경 API에 사용하는 dto들을 users/dto로 이동
