+++ src/docs/dto/user.dto.ts
@@ -1,12 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import {
-  IsString,
-  IsNotEmpty,
-  Length,
-  IsBoolean,
-  IsOptional,
-  Matches,
-} from 'class-validator';
+import { IsString, IsNotEmpty, Length } from 'class-validator';
 
 /**
  * 유저 정보
@@ -99,116 +92,6 @@ export class GetMyInfoResponseDto {
     type: GetMyInfoDataDto,
   })
   data: GetMyInfoDataDto;
-}
-/**
- * 유저 세팅 수정
- */
-export class UpdateUserSettingsRequestDto {
-  @ApiProperty({
-    description: '변경할 닉네임',
-    example: '새로운닉네임',
-    maxLength: 50,
-    required: false,
-  })
-  @IsString()
-  @IsOptional()
-  @Length(1, 50)
-  nickname?: string;
-
-  @ApiProperty({
-    description: '프로필 이미지 종류 (01-05)',
-    example: '03',
-    enum: ['01', '02', '03', '04', '05'],
-    required: false,
-  })
-  @IsString()
-  @IsOptional()
-  @Matches(/^0[1-5]$/, {
-    message: 'profileImageCode는 01부터 05까지의 값이어야 합니다.',
-  })
-  profileImageCode?: string;
-
-  @ApiProperty({
-    description: '예약 알림 수신 설정',
-    example: false,
-    required: false,
-  })
-  @IsBoolean()
-  @IsOptional()
-  reservationAlarmSetting?: boolean;
-
-  @ApiProperty({
-    description: '콕찌르기 알림 수신 설정',
-    example: true,
-    required: false,
-  })
-  @IsBoolean()
-  @IsOptional()
-  kokAlarmSetting?: boolean;
-}
-
-export class UpdateUserSettingsDataDto {
-  @ApiProperty({
-    description: '사용자 ID',
-    example: 123,
-  })
-  userId: number;
-
-  @ApiProperty({
-    description: '디바이스 고유 식별자',
-    example: 'iPhone_ABC123XYZ',
-  })
-  deviceId: string;
-
-  @ApiProperty({
-    description: '사용자 닉네임',
-    example: '새로운닉네임',
-  })
-  nickname: string;
-
-  @ApiProperty({
-    description: '프로필 이미지 파일명',
-    example: 'IMG_001',
-  })
-  profileImageName: string;
-
-  @ApiProperty({
-    description: '예약 알림 수신 설정',
-    example: false,
-  })
-  reservationAlarmSetting: boolean;
-
-  @ApiProperty({
-    description: '콕찌르기 알림 수신 설정',
-    example: true,
-  })
-  kokAlarmSetting: boolean;
-
-  @ApiProperty({
-    description: '설정 변경 시간',
-    example: '2024-06-03T15:30:00Z',
-  })
-  updatedAt: string;
-}
-
-export class UpdateUserSettingsResponseDto {
-  @ApiProperty({
-    description: '응답 코드',
-    example: '200',
-  })
-  code: string;
-
-  @ApiProperty({
-    description: '응답 메시지',
-    example: 'OK',
-  })
-  message: string;
-
-  @ApiProperty({
-    description: '변경된 사용자 설정 정보',
-    type: UpdateUserSettingsDataDto,
-  })
-  data: UpdateUserSettingsDataDto;
 }
 
 export class CheckNicknameQueryDto {
